requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
av==14.3.0
babel==2.17.0
bce-python-sdk==0.9.29
beautifulsoup4==4.13.3
blake3==1.0.4
blinker==1.9.0
boto3==1.37.29
botocore==1.37.29
bottle==0.13.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cn-clip==1.5.1
coloredlogs==15.0.1
common==0.1.2
compressed-tensors==0.9.2
contourpy==1.3.1
cupy-cuda12x==13.4.1
cycler==0.12.1
data==0.4
datasets==3.5.0
decorator==5.2.1
decord==0.6.0
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dual==0.1.0
dynamo3==0.4.10
easydict==1.13
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
Flask==3.1.0
flask-babel==4.0.0
flatbuffers==25.2.10
flywheel==0.5.4
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
funcsigs==1.0.2
future==1.0.0
gast==0.3.3
gdown==5.2.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgviz==1.7.6
importlib_metadata==8.6.1
interegular==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
labelme==5.8.1
lark==1.2.2
lazy_loader==0.4
Levenshtein==0.27.1
llguidance==0.7.11
llvmlite==0.43.0
lm-format-enforcer==0.10.11
lmdb==1.3.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mistral_common==1.5.4
modelscope==1.24.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.3.1
multiprocess==0.70.16
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.60.0
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.21.0
openai==1.70.0
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
osam==0.2.3
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
paddle==1.0.2
paddle-bfloat==0.1.7
paddleclas==2.5.2
paddlepaddle==2.5.0
paddlepaddle-gpu==2.6.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
peewee==3.17.9
peft==0.15.1
pillow==11.1.0
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.1
protobuf==6.30.2
prox==0.0.17
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycountry==24.6.1
pycryptodome==3.22.0
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
PyMySQL==1.1.1
pynvml==12.0.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-geoip-python3==1.3
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
qwen-vl-utils==0.0.10
RapidFuzz==3.12.2
rarfile==4.2
ray==2.44.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
swankit==0.1.7
swanlab==0.5.5
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tight==0.1.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.97
ultralytics-thop==2.0.14
urllib3==2.3.0
uv==0.6.12
uvicorn==0.34.0
uvloop==0.21.0
visualdl==2.5.3
vllm==0.8.2
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xformers==0.0.29.post2
xgrammar==0.1.16
xxhash==3.5.0
yarl==1.18.3
zhconv==1.4.3
zipp==3.21.0
